Split the CheckRules half out of runner_spec.rb
A spec FILE is binpacker's scheduling unit and cannot be split across
workers, so `runner_spec.rb` at 5,955 lines and ~167s was the CI test
matrix's makespan on its own: shard 1 gave it a whole worker while its
other three sat at ~122s, a 24.6% intra-shard deviation against 5.8%
and 4.3% for its siblings. The twelve-worker floor for the suite is
~134s, which shards 2 and 3 hit exactly, and no partition or shard
count can put a 167s file under it.

Attributing the file's per-example timings to its top-level describes
found the cost concentrated rather than spread: `CheckRules
diagnostics (Slice 7 phase 8)` was 65.2% of it (225 of 357 examples,
93.3s of 143.0s locally), and the next block down was 9.4%. It also
sits on no shared state — the whole `analyze` harness lives in
spec/support/runner_helpers.rb — so it moves as a unit.

The block keeps its own `describe` wrapper inside the new file rather
than being flattened to the top level. Nothing is re-indented, so the
squiggly heredocs are untouched, and every example's full name is
unchanged: `--example` filters still resolve and binpacker's timing
history keeps matching its keys. A `--dry-run` of both sides yields
the same 355 full descriptions.

Widen spec_helper's derived-metadata match, which pinned `type:
:runner` — and so `RunnerHelpers#analyze` — to the exact old path.
`runner_pool_spec.rb` still deliberately does not match it.
